Restructure the codes
