+++ lib/main.dart
@@ -1,13 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'package:pse_assignment/screens/plan_for_a_diet_screens/health_data_input_screen.dart';
-import './screens/plan_for_a_diet_screens/health_data_input_screen.dart';
-import 'package:pse_assignment/screens/plan_for_a_diet_screens/plan_for_a_diet_providers/diet_plan_data.dart';
-import './screens/plan_for_a_diet_screens/daily_detail_screen.dart';
-import './screens/plan_for_a_diet_screens/diet_timetable_screen.dart';
-import './screens/plan_for_a_diet_screens/planning_option_screen.dart';
-
 import './screens/health_input_form.dart';
 import './screens/filter_screen.dart';
 import './screens/home_page.dart';
@@ -17,8 +10,15 @@ import './providers/products.dart';
 import 'package:flutter_statusbarcolor/flutter_statusbarcolor.dart';
 import './providers/user_input.dart';
 import './providers/filtered_saved_list.dart';
-import './screens/plan_for_a_diet_screens/plan_for_a_diet_providers/user_health_data.dart';
 import './screens/scan_screen.dart';
+
+// The following are all the files supporting for Plan For A Diet.
+import './plan_for_a_diet/plan_for_a_diet_screens/health_data_input_screen.dart';
+import './plan_for_a_diet/plan_for_a_diet_providers/diet_plan_data.dart';
+import './plan_for_a_diet/plan_for_a_diet_screens/daily_detail_screen.dart';
+import './plan_for_a_diet/plan_for_a_diet_screens/diet_timetable_screen.dart';
+import './plan_for_a_diet/plan_for_a_diet_screens/planning_option_screen.dart';
+import './plan_for_a_diet/plan_for_a_diet_providers/user_health_data.dart';
 
 void main() {
   runApp(MyApp());
