+++ lib/plan_for_a_diet/plan_for_a_diet_providers/daily_data.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../../providers/data_helper.dart';
+import '../../providers/data_helper.dart';
 import 'package:provider/provider.dart';
 import 'user_health_data.dart';
 import 'dart:math';
