+++ lib/plan_for_a_diet/plan_for_a_diet_providers/diet_plan_data.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'user_health_data.dart';
-import '../../../providers/data_helper.dart';
+import '../../providers/data_helper.dart';
 import 'daily_data.dart';
 
 class DietPlanData with ChangeNotifier{
