+++ lib/plan_for_a_diet/plan_for_a_diet_screens/health_data_input_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import './plan_for_a_diet_providers/user_health_data.dart';
-import '../plan_for_a_diet_screens/plan_for_a_diet_providers/diet_plan_data.dart';
-import './plan_for_a_diet_widgets/checkbox_list_widget.dart';
+import '../plan_for_a_diet_providers/user_health_data.dart';
+import '../plan_for_a_diet_providers/diet_plan_data.dart';
+import '../plan_for_a_diet_widgets/checkbox_list_widget.dart';
 
 class HealthDataInputScreen extends StatelessWidget {
   double _userHeight = 0;
