+++ lib/plan_for_a_diet/plan_for_a_diet_screens/planning_option_screen.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../plan_for_a_diet_screens/plan_for_a_diet_providers/diet_plan_data.dart';
+import '../plan_for_a_diet_providers/diet_plan_data.dart';
 
-import '../plan_for_a_diet_screens/plan_for_a_diet_providers/diet_plan_data.dart';
-import './plan_for_a_diet_providers/user_health_data.dart';
+import '../plan_for_a_diet_providers/user_health_data.dart';
 
 class PlanningOptionScreen extends StatelessWidget {
